claudecode/self_reflecting_agent/requirements.txt
# Core framework dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
dspy-ai>=2.4.0

# Memory and storage
mem0ai>=0.1.0
chromadb>=0.5.0
neo4j>=5.0.0

# RAG components
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# Observability and evaluation
mlflow>=2.15.0
prometheus-client>=0.20.0

# Utilities
pydantic>=2.8.0
pyyaml>=6.0.0
rich>=13.0.0
asyncio-throttle>=1.0.0

# Development and testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
isort>=5.13.0
mypy>=1.11.0